src/functions/requirements.txt
azure-functions>=1.21.0
azure-identity>=1.19.0
azure-storage-blob>=12.24.0
httpx>=0.28.0
Pillow>=11.0.0
img2pdf>=0.5.0
fpdf2>=2.8.0
